# Add Regex FindFirstChar optimization for a literal after an atomic loop