--- a/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexFindOptimizations.cs
+++ b/src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexFindOptimizations.cs
@@ -121,12 +121,19 @@ namespace System.Text.RegularExpressions
 
             // We're now left-to-right only and looking for sets.
 
+            // As a backup, see if we can find a literal after a leading atomic loop.  That might be better than whatever sets we find, so
+            // we want to know whether we have one in our pocket before deciding whether to use a leading set.
+            (RegexNode LoopNode, (char Char, string? String, char[]? Chars) Literal)? literalAfterLoop = RegexPrefixAnalyzer.FindLiteralFollowingLeadingAtomicLoop(tree);
+
             // Build up a list of all of the sets that are a fixed distance from the start of the expression.
             List<(char[]? Chars, string Set, int Distance, bool CaseInsensitive)>? fixedDistanceSets = RegexPrefixAnalyzer.FindFixedDistanceSets(tree, culture, thorough: !interpreter);
-            if (fixedDistanceSets is not null)
-            {
-                Debug.Assert(fixedDistanceSets.Count != 0);
+            Debug.Assert(fixedDistanceSets is null || fixedDistanceSets.Count != 0);
 
+            // If we got such sets, we'll likely use them.  However, if the best of them is something that doesn't support a vectorized
+            // search and we did successfully find a literal after an atomic loop we could search instead, we prefer the vectorizable search.
+            if (fixedDistanceSets is not null &&
+                (fixedDistanceSets[0].Chars is not null || literalAfterLoop is null))
+            {
                 // Determine whether to do searching based on one or more sets or on a single literal. Compiled engines
                 // don't need to special-case literals as they already do codegen to create the optimal lookup based on
                 // the set's characteristics.
@@ -162,6 +169,15 @@ namespace System.Text.RegularExpressions
                 }
                 return;
             }
+
+            // If we found a literal we can search for after a leading atomic loop, use it.
+            if (literalAfterLoop is not null)
+            {
+                FindMode = FindNextStartingPositionMode.LiteralAfterAtomicLoop_LeftToRight_CaseSensitive;
+                LiteralAfterAtomicLoop = literalAfterLoop;
+                _asciiLookups = new uint[1][];
+                return;
+            }
         }
 
         /// <summary>Gets the selected mode for performing the next <see cref="TryFindNextStartingPosition"/> operation</summary>
@@ -179,6 +195,9 @@ namespace System.Text.RegularExpressions
         /// <summary>When in fixed distance set mode, gets the set and how far it is from the start of the pattern.</summary>
         /// <remarks>The case-insensitivity of the 0th entry will always match the mode selected, but subsequent entries may not.</remarks>
         public List<(char[]? Chars, string Set, int Distance, bool CaseInsensitive)>? FixedDistanceSets { get; }
+
+        /// <summary>When in literal after atomic loop node, gets the literal to search for and the RegexNode representing the leading loop.</summary>
+        public (RegexNode LoopNode, (char Char, string? String, char[]? Chars) Literal)? LiteralAfterAtomicLoop { get; }
 
         /// <summary>Try to advance to the next starting position that might be a location for a match.</summary>
         /// <param name="textSpan">The text to search.</param>
@@ -601,6 +620,54 @@ namespace System.Text.RegularExpressions
                         return false;
                     }
 
+                // There's a literal after an atomic loop.  Find the literal, then walk backwards through the loop to find the starting position.
+                case FindNextStartingPositionMode.LiteralAfterAtomicLoop_LeftToRight_CaseSensitive:
+                    {
+                        Debug.Assert(LiteralAfterAtomicLoop is not null);
+                        (RegexNode loopNode, (char Char, string? String, char[]? Chars) literal) = LiteralAfterAtomicLoop.GetValueOrDefault();
+
+                        Debug.Assert(loopNode.Type == RegexNode.Setloopatomic);
+                        Debug.Assert(loopNode.N == int.MaxValue);
+
+                        int startingPos = pos;
+                        while (true)
+                        {
+                            ReadOnlySpan<char> slice = textSpan.Slice(startingPos, end - startingPos);
+
+                            // Find the literal.  If we can't find it, we're done searching.
+                            int i = literal.String is not null ? slice.IndexOf(literal.String.AsSpan()) :
+                                    literal.Chars is not null ? slice.IndexOfAny(literal.Chars.AsSpan()) :
+                                    slice.IndexOf(literal.Char);
+                            if (i < 0)
+                            {
+                                break;
+                            }
+
+                            // We found the literal.  Walk backwards from it finding as many matches as we can against the loop.
+                            int prev = i;
+                            while ((uint)--prev < (uint)slice.Length && RegexCharClass.CharInClass(slice[prev], loopNode.Str!, ref _asciiLookups![0])) ;
+
+                            // If we found fewer than needed, loop around to try again.  The loop doesn't overlap with the literal,
+                            // so we can start from after the last place the literal matched.
+                            if ((i - prev - 1) < loopNode.M)
+                            {
+                                startingPos += i + 1;
+                                continue;
+                            }
+
+                            // We have a winner.  The starting position is just after the last position that failed to match the loop.
+                            // TODO: It'd be nice to be able to communicate literalPos as a place the matching engine can start matching
+                            // after the loop, so that it doesn't need to re-match the loop.  This might only be feasible for RegexCompiler
+                            // and the source generator after we refactor them to generate a single Scan method rather than separate
+                            // FindFirstChar / Go methods.
+                            pos = startingPos + prev + 1;
+                            return true;
+                        }
+
+                        pos = end;
+                        return false;
+                    }
+
                 // Nothing special to look for.  Just return true indicating this is a valid position to try to match.
 
                 default:
@@ -659,6 +726,9 @@ namespace System.Text.RegularExpressions
         FixedSets_LeftToRight_CaseSensitive,
         /// <summary>One or more sets at a fixed distance from the start of the pattern.  At least the first set is case-insensitive.</summary>
         FixedSets_LeftToRight_CaseInsensitive,
+
+        /// <summary>A literal after a non-overlapping atomic loop at the start of the pattern.  The literal is case-sensitive.</summary>
+        LiteralAfterAtomicLoop_LeftToRight_CaseSensitive,
 
         /// <summary>Nothing to search for. Nop.</summary>
         NoSearch,