… src/libraries/System.Text.RegularExpressions/src/System/Text/RegularExpressions/RegexPrefixAnalyzer.cs
… src/libraries/System.Text.RegularExpressions/tests/Regex.Match.Tests.cs