Fixes https://github.com/dotnet/runtime/issues/62445

If after all of our analysis we can't find anything to IndexOf{Any} as part of finding the next match location, we'll also look for a literal (a single char, a string of chars, or a set of a small number of chars) following an unbounded atomic set loop that starts the pattern.  In that case, we can IndexOf for the literal, and then walk backwards through the characters prior to it looking for the beginning of the match.

The pros / cons of this depend on what you're using Regex for.  For cons, if you're just trying to match against a string to say "is this a blah", this addition is likely going to be slower, as each of the characters that match the starting loop will end up being evaluated twice, once backward as part of finding the starting position and once forward as part of doing the match.  That cost could be mitigated once we have a single Scan method rather than two separate FindFirstChar / Go methods, such that we can more easily pass additional information from FindFirstChar to Go (e.g. not just the starting position, but also where the loop ended).  It could also potentially be mitigated by [#invertscan](https://github.com/dotnet/runtime/issues/62443), at least for some cases.

On the pro side however, if you're using Regex to search for something in some larger body of text, this can have a massive upside, doing a vectorized search for one or more characters rather than walking each and evaluating each against a set.

These pros and cons can be seen with these perf tests:
https://github.com/dotnet/performance/blob/824857deee952aa5496729ca79ed34d7c4eb66d9/src/benchmarks/micro/libraries/System.Text.RegularExpressions/Perf.Regex.Common.cs#L51-L52
https://github.com/dotnet/performance/blob/824857deee952aa5496729ca79ed34d7c4eb66d9/src/benchmarks/micro/libraries/System.Text.RegularExpressions/Perf.Regex.Industry.cs#L75-L76

|     Method |         Toolchain |              Options |              Mean | Ratio |
|----------- |------------------ |--------------------- |------------------:|------:|
|    IsMatch | \main\corerun.exe |                 None |         274.15 ns |  1.00 |
|    IsMatch |   \pr\corerun.exe |                 None |         296.11 ns |  1.03 |
|            |                   |                      |                   |       |
| IsNotMatch | \main\corerun.exe |                 None |         298.84 ns |  1.00 |
| IsNotMatch |   \pr\corerun.exe |                 None |         321.58 ns |  1.08 |
|            |                   |                      |                   |       |
|      Count | \main\corerun.exe |                 None |  87,861,050.35 ns |  1.00 |
|      Count |   \pr\corerun.exe |                 None |   3,745,879.58 ns |  0.04 |
|            |                   |                      |                   |       |
|    IsMatch | \main\corerun.exe |             Compiled |          84.58 ns |  1.00 |
|    IsMatch |   \pr\corerun.exe |             Compiled |         111.19 ns |  1.32 |
|            |                   |                      |                   |       |
| IsNotMatch | \main\corerun.exe |             Compiled |         108.16 ns |  1.00 |
| IsNotMatch |   \pr\corerun.exe |             Compiled |         128.68 ns |  1.20 |
|            |                   |                      |                   |       |
|      Count | \main\corerun.exe |             Compiled |  44,019,520.24 ns |  1.00 |
|      Count |   \pr\corerun.exe |             Compiled |   3,131,700.00 ns |  0.07 |

I think it's worthwhile to take it, but I'm open to feedback to the contrary.